+++ course/src/main/java/in/testpress/course/ui/ContentListFragment.kt
@@ -1,6 +1,7 @@
 package `in`.testpress.course.ui
 
 import `in`.testpress.course.R
+import `in`.testpress.course.TestpressCourse
 import `in`.testpress.course.domain.DomainContent
 import `in`.testpress.course.fragments.BaseContentListFragment
 import `in`.testpress.enums.Status
@@ -13,19 +14,47 @@ import android.view.View
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
 
 class ContentListFragment : BaseContentListFragment(), EmptyViewListener {
 
+    private var contentsURL: String = ""
+    private var chapterId: Long = -1
+    private var productSlug: String? = null
     private lateinit var viewModel: ContentsListViewModel
     private lateinit var mAdapter: ContentListAdapter
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        parseArguments()
+        initializeViewModel()
+    }
+
+    private fun parseArguments() {
+        contentsURL = arguments!!.getString(TestpressCourse.CONTENTS_URL_FRAG) ?: ""
+        chapterId = arguments!!.getLong(TestpressCourse.CHAPTER_ID)
+        productSlug = arguments!!.getString(TestpressCourse.PRODUCT_SLUG)
+    }
+
+    private fun initializeViewModel() {
+        viewModel = ViewModelProvider(this, object : ViewModelProvider.Factory {
+            override fun <T : ViewModel> create(modelClass: Class<T>): T {
+                return ContentsListViewModel(ContentsRepository(requireContext(), chapterId)) as T
+            }
+        }).get(ContentsListViewModel::class.java)
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         mAdapter = ContentListAdapter(chapterId, productSlug)
-        recyclerView.adapter = mAdapter
+        recyclerView.apply {
+            layoutManager = LinearLayoutManager(requireContext())
+            adapter = mAdapter
+            addItemDecoration(DividerItemDecoration(context, DividerItemDecoration.VERTICAL))
+        }
         initializeObservers()
         viewModel.loadContents()
-        swipeRefreshLayout.setOnRefreshListener { viewModel.loadContents() }
     }
 
     private fun initializeObservers() {
@@ -44,7 +73,6 @@ class ContentListFragment : BaseContentListFragment(), EmptyViewListener {
                     if (items.isEmpty()) showEmptyList(resources.getString(R.string.testpress_no_content))
                     mAdapter.contents = items
                     mAdapter.notifyDataSetChanged()
-                    swipeRefreshLayout.isRefreshing = false
                 }
                 Status.ERROR -> {
                     Log.d("ContentListFragment", "Got status ERROR")
@@ -60,20 +88,8 @@ class ContentListFragment : BaseContentListFragment(), EmptyViewListener {
         })
     }
 
-    override fun initializeViewModel() {
-        viewModel = ViewModelProvider(this, object : ViewModelProvider.Factory {
-            override fun <T : ViewModel> create(modelClass: Class<T>): T {
-                return ContentsListViewModel(ContentsRepository(requireContext(), chapterId)) as T
-            }
-        }).get(ContentsListViewModel::class.java)
-    }
-
     override fun onRetryClick() {
         Log.d("onRetryClick", "viewModel load contents")
         viewModel.loadContents()
-    }
-
-    override fun fetchMore() {
-
     }
 }
